backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
python-dotenv>=1.0
python-multipart>=0.0.6
requests>=2.28
google-generativeai>=0.3
transformers>=4.40
openai-whisper>=20230314
beautifulsoup4>=4.12
lxml>=5.2
